requirements.txt
pytorch>=1.2
cloudpickle